Add -fix_names to apply naming fixes suggested by the linter
- Provide suggestions for ALL_CAPS cases
- add -only flag to limit linter to only naming checks

This commit depends on a change in the gorename refactor tool which
changes its return value to include the filenames changed after its
exection.

Note: this is actually a best effort change to automatically refactor a
project that has not been following Go's naming conventions. Although it
fixes most of the naming problems, manual intervention is required when
gorename cannot apply its suggested fixes.